Added support for staff account creation
Added STAFF variable to settings, which is a list of account names
that should be (re)created and added to the staff and sudo groups.

For each staff account, if the local file keys/[username].pub exists,
it will be deployed as ~username/.ssh/authorized_keys on the remote hosts.