Add basic postfix installation task; fix #11
The task install_postfix() will install and (re)configure postfix as either an internet site or a satellite, depending on whether the relay parameter is set, or SMTP_RELAY is defined in settings.

There is plenty of opportunity for enhancements here, since this is just a default postfix install, though it is possibly appropriate to make these enhancements in downstream projects where the postfix configs can be managed via templates, rather than additional customized support baked into cotton.